BIOSTD-24: Submission Files In Separate Collection
- Create a post-processing service class to contain all the post-processing related operations there
- Create a separate post processing test verifying all the required steps are completed
- Rename a couple of methods to clarify their purpose
