+++ submission/submission-core/src/main/kotlin/ac/uk/ebi/biostd/submission/stats/SubmissionStatsCalculator.kt
@@ -4,14 +4,9 @@ import ac.uk.ebi.biostd.persistence.common.model.SubmissionStat
 import ac.uk.ebi.biostd.persistence.common.model.SubmissionStatType.DIRECTORIES
 import ac.uk.ebi.biostd.persistence.common.model.SubmissionStatType.FILES_SIZE
 import ac.uk.ebi.biostd.persistence.common.model.SubmissionStatType.NON_DECLARED_FILES_DIRECTORIES
-import ac.uk.ebi.biostd.persistence.filesystem.api.FileStorageService
-import ebi.ac.uk.extended.model.ExtFile
 import ebi.ac.uk.extended.model.ExtFileType
 import ebi.ac.uk.extended.model.ExtSubmission
 import ebi.ac.uk.extended.model.PersistedExtFile
-import ebi.ac.uk.extended.model.allPageTabFiles
-import ebi.ac.uk.paths.SubmissionFolderResolver
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.filterIsInstance
 import kotlinx.coroutines.flow.firstOrNull
@@ -23,13 +18,8 @@ private val logger = KotlinLogging.logger {}
 
 class SubmissionStatsCalculator(
     private val serializationService: ExtSerializationService,
-    private val fileStorageService: FileStorageService,
-    private val subFolderResolver: SubmissionFolderResolver,
 ) {
-    internal suspend fun calculateStats(sub: ExtSubmission): List<SubmissionStat> {
-        copyPageTabFiles(sub)
-        return calculatePlainStats(sub)
-    }
+    internal suspend fun calculateStats(sub: ExtSubmission): List<SubmissionStat> = calculatePlainStats(sub)
 
     private suspend fun calculatePlainStats(sub: ExtSubmission): List<SubmissionStat> {
         logger.info { "Calculating stats for submission ${sub.accNo}, version ${sub.version}" }
@@ -62,12 +52,4 @@ class SubmissionStatsCalculator(
             .filterIsInstance<PersistedExtFile>()
             .filter { it.type == ExtFileType.FILE }
             .firstOrNull { it.filePath.contains(directoryPath) } != null
-
-    private suspend fun copyPageTabFiles(sub: ExtSubmission): List<ExtFile> {
-        logger.info { "Copying pagetab files for submission ${sub.accNo}, version ${sub.version}" }
-        return with(Dispatchers.IO) {
-            val expectedPath = subFolderResolver.getCopyPageTabPath(sub)
-            sub.allPageTabFiles.onEach { fileStorageService.copyFile(it, expectedPath) }
-        }
-    }
 }
